refactor: fold orchestrate into the solver module
The result types (SolveResult, BatchSolveResult) lived in orchestrate.rs
but were produced by Solver::solve one layer below, so solver.rs reached
*up* into orchestrate for its own return types while orchestrate called
*down* into solver — a module cycle. Move the result types and the thin
solve/solve_batch wrappers next to the Solver that produces and backs
them, leaving a single cohesive solve-API module and deleting the
pure-satellite orchestrate module.

Crate-root paths (within::solve, within::SolveResult, ...) are unchanged,
so within-py and downstream consumers are unaffected.
